Packages
Had to update the project to Unity 2018.03.0f2 in order to edit.
Added a couple of packages I've made,  Face Generator - Add the prefab facelayer libraries to the scene, the face display and put the testing script somewhere in the scene and it will generate a new face every time you hit space.
The inventory box needs to be added to a panel in the UI, then once you specify the number of slots you want it should generate an inventory. If you add the testing script it will add a couple of items to play around with.