Seeding and witches
Modified the hex engine to generate a seeded map.
Added a witch base to the world on start up, this wich base generates 3 witches and if you click on it you can see their image, the idea being you could then select the witch you want and give them an order. Also uses the name generator from Driftwood but you can only see that in the inspector because I haven't added anything to the UI yet to show that.