WitchBaseMenu
Added the ability to select a witch from the base, tell it to move and have the game create a witch in the game world. The witch unit then takes on the colours of the witch's portrait. Doesn't do anything once its in the gameworld, but we can then add scripts to move it around, explore places and fight stuff.